+++ src/frontend/magnify/.storybook/theme.js
@@ -5,5 +5,22 @@ export default {
       size: '14px',
       height: '20px',
     },
+    colors: {
+      brand: '#035ccd',
+      'neutral-1': '#294c79',
+      'light-2': '#f2f7fd',
+      'light-3': '#e5eefa',
+      'accent-1': '#ff0000',
+    },
+    focus: {
+      border: {
+        color: 'light-3',
+      },
+    },
+  },
+  button: {
+    border: {
+      radius: '5px',
+    },
   },
 };