+++ src/frontend/magnify/src/components/index.ts
@@ -1,1 +1,3 @@
 export { default as TestButton, TestButtonVariant } from './TestButton';
+
+export * from './groups';