+++ src/frontend/magnify/src/types/group.interface.ts
@@ -1,0 +1,7 @@
+import { Member } from './member.interface';
+
+export interface Group {
+  id: string;
+  name: string;
+  members: Member[];
+}